requirements.txt
flask
google-generativeai
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
faiss-cpu
pymongo
python-dotenv
nest-asyncio
motor
flask-cors

